+++ app/controllers/carto/api/user_presenter.rb
@@ -1,17 +1,22 @@
-class Carto::Api::UserPresenter
 
-  def initialize(user)
-    @user = user
+module Carto
+  module Api
+    class UserPresenter
+
+      def initialize(user)
+        @user = user
+      end
+
+      def to_poro
+        return {} if @user.nil?
+        {
+            id:         @user.id,
+            username:   @user.username,
+            avatar_url: @user.avatar_url,
+            base_url:   @user.public_url
+        }
+      end
+
+    end
   end
-
-  def to_poro
-    return {} if @user.nil?
-    {
-        id:         @user.id,
-        username:   @user.username,
-        avatar_url: @user.avatar_url,
-        base_url:   @user.public_url
-    }
-  end
-
 end