+++ app/helpers/application_helper.rb
@@ -61,11 +61,6 @@ module ApplicationHelper
       end
   end
 
-  def default_fallback_basemap
-    Cartodb.config[:basemaps].present? and Cartodb.config[:basemaps]['default'].present? ? Cartodb.config[:basemaps]['default'] : {}
-  end
-
-  module_function :default_fallback_basemap
   module_function :maps_api_template
   module_function :sql_api_template
 
@@ -88,8 +83,7 @@ module ApplicationHelper
       tumblr_api_key:             Cartodb.config[:tumblr]['api_key'],
       max_asset_file_size:        Cartodb.config[:assets]["max_file_size"],
       watcher_ttl:                Cartodb.config[:watcher].try("fetch", 'ttl', 60),
-      default_fallback_basemap:   default_fallback_basemap,
-      upgrade_url:                cartodb_com_hosted? ? "''" : "#{current_user.upgrade_url(request.protocol)}" 
+      upgrade_url:                cartodb_com_hosted? ? "''" : "#{current_user.upgrade_url(request.protocol)}"
     }
 
     if Cartodb.config[:datasource_search].present? && Cartodb.config[:datasource_search]['twitter_search'].present? \